requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
Django==3.2.7
django-filter==21.1
djangorestframework==3.12.4
idna==3.2
importlib-metadata==4.8.1
Markdown==3.3.4
nose==1.3.7
pkg-resources==0.0.0
pytz==2021.3
PyYAML==5.4.1
pyyml==0.0.2
Random-Word==1.0.7
requests==2.26.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
urllib3==1.26.7
zipp==3.6.0
